Support external specified avro tools jar
This is helpful when you already pull down an avro
version from the non-standard maven repo and you
just want to use that in the toolchain. Helps ensure
that you have the same version everywhere in your build,
rather than using a maybe close, but not exactly the same,
version from maven central.